Time Shifting Charts
All charts ready and correctly displaying. They start with Afghanistan and an indicator that has no data, so that might be advisable to change. Also, added minor functionalities to norm.py and shift_corr.py, and started planning the test for all methods used in the time shift notebook (still pending the test itself).